+ftp +dmx +mqtt

+++ main/modbus.h
@@ -14,5 +14,5 @@ esp_err_t setHoldingQueue(uint8_t slaveid, uint8_t adr, uint16_t value);
 esp_err_t mbInit();
 void setReadTimeOut(uint16_t val);
 void processQueue();
-esp_err_t executeModbusCommand(uint8_t slaveId, uint8_t command, uint8_t start, uint8_t qty, uint8_t **response);
+esp_err_t executeModbusCommand(uint8_t slaveId, uint8_t command, uint16_t start, uint8_t qty, uint8_t **response);
 void test();
